% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl_helper_functions.R
\name{calculate_si_probability_matrix}
\alias{calculate_si_probability_matrix}
\title{Calculate Serial Interval Probability Matrix}
\usage{
calculate_si_probability_matrix(day_diffs, si_mean, si_sd, si_dist)
}
\arguments{
\item{day_diffs}{Matrix of day differences between each pair of cases}

\item{si_mean}{Mean of the serial interval distribution}

\item{si_sd}{Standard deviation of the serial interval distribution}

\item{si_dist}{Distribution type ("gamma" or "normal")}
}
\value{
Matrix of serial interval probabilities
}
\description{
Creates a matrix of probabilities based on the serial interval distribution
for all pairwise day differences.
}
