% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl_helper_functions.R
\name{calculate_truncation_correction}
\alias{calculate_truncation_correction}
\title{Calculate Right Truncation Correction}
\usage{
calculate_truncation_correction(dates, si_mean, si_sd, si_dist)
}
\arguments{
\item{dates}{Vector of dates}

\item{si_mean}{Mean of the serial interval}

\item{si_sd}{Standard deviation of the serial interval}

\item{si_dist}{Distribution type ("gamma" or "normal")}
}
\value{
Vector of correction factors
}
\description{
Calculates correction factors for right truncation of the time series.
}
