% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{check_args}
\alias{check_args}
\title{Function to check the arguments of \code{simulate_seir_ode()}.}
\usage{
check_args(arnaught, t_E, t_I, N, S_init, E_init, I_init, n_t, n_steps_per_t)
}
\arguments{
\item{arnaught}{Basic reproduction number (ratio), squiggly-R-0. Average number of new infections produced by an infection in a susceptible population. A scalar or a vector of length \code{n_t + 1}, which specifies R0 at the start of each timestep. R0 is linearly interpolated between timesteps.}

\item{t_E}{numeric; Mean latent period. If set to 0, the model reduces to an SIR.}

\item{t_I}{numeric; Mean duration of infectiousness.}

\item{N}{integer; Total population size.}

\item{S_init}{integer; number of individuals starting in the susceptible state.}

\item{E_init}{integer; number of individuals starting in the exposed state.}

\item{I_init}{interger; number of individuals starting in the infectious state.}

\item{n_t}{integer; Number of units of time to simulate.}

\item{n_steps_per_t}{integer; number of time steps, defaults to 1.}
}
\value{
error message
}
\description{
The function checks that the initial values for the different states are whole numbers and that $R_0$, $t_E$ and $t_I$ are numeric.
}
