% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{construct_beta}
\alias{construct_beta}
\title{Function to construct the transmission rate from $R_0$, $t_I$ (mean infectious period), and $n_t$.}
\usage{
construct_beta(arnaught, t_I, n_t)
}
\arguments{
\item{arnaught}{Basic reproduction number (ratio), squiggly-R-0. Average number of new infections produced by an infection in a susceptible population. A scalar or a vector of length \code{n_t + 1}, which specifies R0 at the start of each timestep. R0 is linearly interpolated between timesteps.}

\item{t_I}{numeric; Mean duration of infectiousness.}

\item{n_t}{integer; Number of units of time to simulate.}
}
\value{
vector of transmission rates for each time point
}
\description{
Function to construct the transmission rate from $R_0$, $t_I$ (mean infectious period), and $n_t$.
}
