% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_tri_dist.R
\name{conv_tri_dist}
\alias{conv_tri_dist}
\title{Convolution of the triangular distribution with the mixture component density (continuous case)}
\usage{
conv_tri_dist(x, sigma, r, mu, route)
}
\arguments{
\item{x}{description}

\item{sigma}{description}

\item{r}{description}

\item{mu}{description}

\item{route}{integer; between 1 and 7 and indicates the route of transmission.}
}
\description{
We split the folded normal distribution for Primary-Secondary, Primary-Tertiary and Primary-Quaternary routes into two parts
\itemize{
\item component 1: Co-Primary route
\item component 2+3: Primary-Secondary route
\item component 4+5: Primary-Tertiary route
\item component 6+7: Primary-Quaternary route
}
}
