% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abs_prob.R
\name{get_abs_prob}
\alias{get_abs_prob}
\title{Calculate the absolute probability of an index case to case interval belonging to a transmission route component}
\usage{
get_abs_prob(dat)
}
\arguments{
\item{dat}{a (numeric) vector of index case to case intervals}
}
\value{
matrix where each row represents a transmission route and each column represents the absolue probability of belonging to that transmission route for each index case to case interval in dat
}
\description{
Calculate the absolute probability of an index case to case interval belonging to a transmission route component
}
\examples{
my_data<-c(rep(1,38),rep(2,39),rep(3,30),rep(4,17),rep(5,7))
get_abs_prob(my_data)
}
