% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseR_functions.R
\name{integrate_Rt}
\alias{integrate_Rt}
\title{Calculate the instantaneous case reproductive number using the deterministic SEIR output}
\usage{
integrate_Rt(
  beta_t,
  sigma,
  gamma,
  N,
  T_final,
  E0,
  I0,
  dt_int = 0.1,
  dt_Rt = 1,
  T_inf = 1000
)
}
\arguments{
\item{beta_t}{Function that returns beta (t) for any time t.}

\item{sigma}{Rate of leaving latent class.}

\item{gamma}{Rate of leaving infected class.}

\item{N}{Total population size.}

\item{T_final}{Final time.}

\item{E0}{Initial number of individuals in exposed state.}

\item{I0}{Initial number of individuals in infectious state.}

\item{dt_int}{Integration time step; defaults to \code{dt_int = 0.1}.}

\item{dt_Rt}{Case reproduction number output time step; defaults to \code{dt_Rt = 1}.}

\item{T_inf}{Final time in SEIR integration. Should be at least one generation interval higher than T_final.}
}
\value{
Data frame with estimated case reproduction number for evert time point.
}
\description{
Calculate the instantaneous case reproductive number using the deterministic SEIR output
}
