% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseR_functions.R
\name{integrate_seir}
\alias{integrate_seir}
\title{Function to integrate SEIR}
\usage{
integrate_seir(beta_t, sigma, gamma, N, T_final, E0, I0, dt = 0.01)
}
\arguments{
\item{beta_t}{Function that returns beta (t) for any time t.}

\item{sigma}{Rate of leaving latent class.}

\item{gamma}{Rate of leaving infected class.}

\item{N}{Total population size.}

\item{T_final}{Final time step.}

\item{E0}{Initial number of individuals in exposed state.}

\item{I0}{Initial number of individuals in infectious state.}

\item{dt}{Size of time step over which to integrate SEIR; defaults to \code{dt = 0.01}.}
}
\value{
Data frame with number of individuals in each state for each time point.
}
\description{
Function to integrate SEIR
}
