% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Function to check if input number is an integer}
\usage{
is.wholenumber(x)
}
\arguments{
\item{x}{numeric; value to check whether it is a whole number}
}
\value{
logical expression; returns TRUE if x is a whole number
}
\description{
Function to check if input number is an integer
}
