% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseR_functions.R
\name{make_p_infectious}
\alias{make_p_infectious}
\title{Function to compute the probability, $p(u)$, that an individual is infectious u days after the moment of infection.}
\usage{
make_p_infectious(sigma, gamma)
}
\arguments{
\item{sigma}{rate of leaving compartment E}

\item{gamma}{rate of leaving compartment I}
}
\value{
a function that computes the probability of infectiousnessu days after the moment of infection
}
\description{
Function to compute the probability, $p(u)$, that an individual is infectious u days after the moment of infection.
}
