% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_estim.R
\name{rt_estim}
\alias{rt_estim}
\title{Estimate Time-varying Reproduction Number}
\usage{
rt_estim(inc_dat, mean_si, sd_si, dist_si = "normal", perturb_si_dist = FALSE)
}
\arguments{
\item{inc_dat}{data frame; data frame with incidence data. The data frame should have two columns: inc (daily incidence) and onset_date (date of onset of symptoms).}

\item{mean_si}{numeric; mean of serial interval distribution}

\item{sd_si}{numeric; standard deviation of serial interval distribution}

\item{dist_si}{string; distribution to be assumed for serial interval. Accepts "normal" or "gamma".}

\item{perturb_si_dist}{logical; if true a perturbed serial interval is generated based on dist_si, mean_si, and sd_si, then the mean and sd of the perturbed distribution are used in the calculation of the likelihood function.}
}
\value{
A numeric vector with the expected reproduction number for each day.
}
\description{
This function estimates the time-varying case reproduction number (R_t) using the method proposed by Wallinga and Lipsitch (Proc. Biol. Sci. 2007).
The function takes as input a data frame with incidence data and requires the specification of the serial interval distribution. Note, We assume the average generation interval (time between infections of infector and infectee) is equal to the average serial interval (time between symptom onsets of infector and infectee).
The user must input the mean and standard deviation of the serial interval distribution and select the functional form: Normal or Gamma.
}
