% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_estim_w_boot.R
\name{rt_estim_w_boot}
\alias{rt_estim_w_boot}
\title{Estimate Time-varying Reproduction Number}
\usage{
rt_estim_w_boot(inc_dat, mean_si, sd_si, dist_si = "normal", n_bootstrap = 100)
}
\arguments{
\item{inc_dat}{data frame; data frame with incidence data. The data frame should have two columns: inc (daily incidence) and onset_date. Onset_date does not have to be in date format, it can be a column of days, such as c(1,2,3,4, ... ),}

\item{mean_si}{numeric; mean of serial interval distribution}

\item{sd_si}{numeric; standard deviation of serial interval distribution}

\item{dist_si}{string; distribution to be assumed for serial interval. Accepts "normal" or "gamma".}

\item{n_bootstrap}{integer; number of bootstrap samples of the serial interval distribution}
}
\value{
A named list of two data frames. The first data frame (\code{results}) contains the mean, mediean, 2.5th percentile, and 97.5th percentile from all of the boot strapped samples. The second data frame (\code{boot_samples}) contains all boot strap samples for each time point.
}
\description{
This function estimates the time-varying reproduction number (R_t) using the method proposed by Wallinga and Teunis (AJE 2004).
The function takes as input a data frame with incidence data and requires the specification of the serial interval distribution.
The user must input the mean and standard deviation of the serial interval distribution and select the functional form: Normal or Gamma.
}
