% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_seir_ode.R
\name{simulate_seir_ode}
\alias{simulate_seir_ode}
\title{Simulate a deterministic ODE approximation of a continuous-time, discrete-state stochastic S(E)IR model.}
\usage{
simulate_seir_ode(
  arnaught,
  t_E,
  t_I,
  N,
  S_init,
  E_init,
  I_init,
  n_t,
  n_steps_per_t = 1
)
}
\arguments{
\item{arnaught}{Basic reproduction number (ratio), squiggly-R-0. Average number of new infections produced by an infection in a susceptible population. A scalar or a vector of length \code{n_t + 1}, which specifies R0 at the start of each timestep. R0 is linearly interpolated between timesteps.}

\item{t_E}{Mean latent period. If set to 0, the model reduces to an SIR.}

\item{t_I}{Mean duration of infectiousness.}

\item{N}{integer; Total population size.}

\item{S_init}{integer; number of individuals starting in the susceptible state.}

\item{E_init}{integer; number of individuals starting in the exposed state.}

\item{I_init}{interger; number of individuals starting in the infectious state.}

\item{n_t}{Number of units of time to simulate.}

\item{n_steps_per_t}{integer; number of time steps, defaults to 1.}
}
\value{
data frame with number of individuals transitioning between each state per time step
}
\description{
Original code created by Ed Baskerville (15 April 2020)
}
\details{
This model has no age structure.
}
