% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl_helper_functions.R
\name{smooth_estimates}
\alias{smooth_estimates}
\title{Apply Smoothing to Estimates}
\usage{
smooth_estimates(r_estimate, window)
}
\arguments{
\item{r_estimate}{Vector of R estimates}

\item{window}{Smoothing window size}
}
\value{
Smoothed R estimates
}
\description{
Applies moving average smoothing to R estimates.
}
