% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_int.R
\name{vec_int}
\alias{vec_int}
\title{Vectorized integration}
\usage{
vec_int(x, l, u, rt, q)
}
\arguments{
\item{x}{a (numeric) vector of index case to case intervals}

\item{l}{lower bound for integration}

\item{u}{upper bound for integration}

\item{rt}{route of transmission; integer from 1 to 7}

\item{q}{character string; quantity to estimate, "zero", "lower", "upper"}
}
\value{
a vector of solutions for the integral for every value of x
}
\description{
Vectorized integration
}
\examples{
N <- 1000; hmu<-15; hsigma<-3
set.seed(1234)
CP <- fdrtool::rhalfnorm((0.2*N),theta=sqrt(pi/2)/(sqrt(2)*hsigma))
PS <- stats::rnorm(0.5*N,mean=hmu,sd=hsigma)
PT <- stats::rnorm(0.3*N,mean=2*hmu,sd=sqrt(2)*hsigma)
PQ <- stats::rnorm(0.1*N,mean=3*hmu,sd=sqrt(3)*hsigma)
icc_intervals <- round(c(CP,PS,PT,PQ))

tmp <- vec_int(icc_intervals, l = icc_intervals, u = (icc_intervals + 1), rt = 1, q = "zero")
}
