% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_var.R
\name{weighted.var}
\alias{weighted.var}
\title{Calculate weighted variance}
\usage{
weighted.var(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{w}{numeric vecotr of weights to be used in the caluclation of the weighted variance of x}

\item{na.rm}{logical; remove NA values?}
}
\value{
numeric estimate of the weighted variance of x
}
\description{
This function calculates the weighted variance which is then used in the estimation step (E-step) of the EM algorithm
}
\examples{
a <- 1:10
b <- eq(0.1, 1, length.out = 10)

weighted.var(a, b)
}
