% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_likelihood.R
\name{wt_loglik}
\alias{wt_loglik}
\title{weighted likelihood
each point adds to likelihood given weight belonging to component 2: w2}
\usage{
wt_loglik(dat, par, tau2)
}
\arguments{
\item{dat}{data to be used for maximisation}

\item{par}{parameters to be maximised}

\item{tau2}{numeric; mixture weights for (primary, secondary) pairs}
}
\value{
value of the negative sum of the weighted log likelihood
}
\description{
weighted likelihood
each point adds to likelihood given weight belonging to component 2: w2
}
