% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl_helper_functions.R
\name{calculate_bootstrap_ci}
\alias{calculate_bootstrap_ci}
\title{Calculate Bootstrap Confidence Intervals}
\usage{
calculate_bootstrap_ci(
  incidence,
  si_prob,
  dates,
  si_mean,
  si_sd,
  si_dist,
  smoothing,
  n_bootstrap,
  conf_level
)
}
\arguments{
\item{incidence}{Vector of case counts}

\item{si_prob}{Matrix of serial interval probabilities}

\item{dates}{Vector of dates}

\item{si_mean}{Mean of the serial interval}

\item{si_sd}{Standard deviation of the serial interval}

\item{si_dist}{Distribution type}

\item{smoothing}{Smoothing window size}

\item{n_bootstrap}{Number of bootstrap samples}

\item{conf_level}{Confidence level}
}
\value{
List with lower and upper bounds for R and R_corrected
}
\description{
Calculates confidence intervals using bootstrap resampling.
}
