% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl_helper_functions.R
\name{calculate_r_estimates}
\alias{calculate_r_estimates}
\title{Calculate R From Incidence and Serial Interval}
\usage{
calculate_r_estimates(
  incidence,
  si_prob,
  dates,
  si_mean,
  si_sd,
  si_dist,
  smoothing
)
}
\arguments{
\item{incidence}{Vector of case counts}

\item{si_prob}{Matrix of serial interval probabilities}

\item{dates}{Vector of dates}

\item{si_mean}{Mean of the serial interval}

\item{si_sd}{Standard deviation of the serial interval}

\item{si_dist}{Distribution type}

\item{smoothing}{Smoothing window size}
}
\value{
List with R and R_corrected vectors
}
\description{
Core function to calculate R estimates from incidence data and serial interval probabilities.
}
