% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl_helper_functions.R
\name{create_day_diff_matrix}
\alias{create_day_diff_matrix}
\title{Create Day Difference Matrix}
\usage{
create_day_diff_matrix(dates)
}
\arguments{
\item{dates}{Vector of dates for each case}
}
\value{
Matrix of day differences
}
\description{
Creates a matrix of day differences between all pairs of cases.
}
