% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f0.R
\name{f0}
\alias{f0}
\title{Calculate f0 for Different Components}
\usage{
f0(x, mu, sigma, comp, dist = "normal")
}
\arguments{
\item{x}{numeric; the value at which to evaluate the function.}

\item{mu}{numeric; the mean value.}

\item{sigma}{numeric; the standard deviation.}

\item{comp}{integer; the component number. Component 1 is Co-Primary. Even components
2i are positive routes, odd components 2i+1 are negative routes (normal only).}

\item{dist}{string; assumed distribution of the serial interval; takes "normal" or "gamma"; defaults to "normal"}
}
\value{
The calculated value of f0.
}
\description{
This function calculates the value of f0 based on the component, where the components
represent the transmission routes: Co-Primary (CP), Primary-Secondary (PS), Primary-Tertiary (PT), and Primary-Quaternary (PQ) and beyond. We split the PS, PT, PQ (and higher) routes into two parts, such that
\itemize{
\item component 1: CP route
\item components 2+3: PS route
\item components 4+5: PT route
\item components 2i + (2i+1): route i+1
}
}
\details{
If the dist = gamma, then the mean (\eqn{\mu}) and standard deviation (sigma) are converted into the
shape (k) and scale (theta) parameters for the gamma distribution, such that the mean (\eqn{\mu}
) and variance (\eqn{\sigma^2}) are given by:
\deqn{\mu = k \times \theta}
\deqn{\sigma^2 = k \times \theta^2}.
}
\examples{
\dontrun{
# Basic example with normal distribution
f0(x = 0.5, mu = 12, sigma = 3, comp = 2, dist = "normal")

# Same parameters with gamma distribution
f0(x = 0.5, mu = 12, sigma = 3, comp = 2, dist = "gamma")
}
}
\keyword{internal}
