% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f0.R
\name{f0}
\alias{f0}
\title{Calculate f0 for Different Components}
\usage{
f0(x, mu, sigma, comp)
}
\arguments{
\item{x}{The value at which to evaluate the function.}

\item{mu}{The mean value.}

\item{sigma}{The standard deviation.}

\item{comp}{The component number (1 to 7).}
}
\value{
The calculated value of f0.
}
\description{
This function calculates the value of f0 based on the component.
}
