% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_norm.R
\name{f_norm}
\alias{f_norm}
\title{Helper function to plot the fit of the serial interval distribution assuming an underlying normal distribution}
\usage{
f_norm(x, w1, w2, w3, mu, sigma)
}
\arguments{
\item{x}{quantile}

\item{w1}{probability weight of being a co-primary case}

\item{w2}{probability weight of being a primary-secondary case}

\item{w3}{probability weight of being a primary-tertiary case}

\item{mu}{mean serial interval}

\item{sigma}{standard deviation of serial interval}
}
\value{
the weighted density for each value of x
}
\description{
Helper function to plot the fit of the serial interval distribution assuming an underlying normal distribution
}
