% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fupper.R
\name{fupper}
\alias{fupper}
\title{Calculate fupper for Different Components}
\usage{
fupper(x, r, mu, sigma, comp, dist = "normal")
}
\arguments{
\item{x}{The value at which to evaluate the function.}

\item{r}{The value of r.}

\item{mu}{The mean value.}

\item{sigma}{The standard deviation.}

\item{comp}{The component number (1 to 7).}

\item{dist}{string; assumed distribution of the serial interval; accepts "normal" or "gamma"; defaults to "normal".}
}
\value{
The calculated value of fupper.
}
\description{
This function calculates the value of fupper based on the component.
}
