% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl_helper_functions.R
\name{generate_case_bootstrap}
\alias{generate_case_bootstrap}
\title{Generate Case Bootstrap Sample}
\usage{
generate_case_bootstrap(incidence)
}
\arguments{
\item{incidence}{Vector of case counts}
}
\value{
Bootstrapped incidence vector
}
\description{
Generates a bootstrap sample of the incidence data by resampling individual cases.
}
