% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_likelihood.R
\name{get_likelihood}
\alias{get_likelihood}
\title{Compute the likelihood of a transmission pair given the serial interval distribution}
\usage{
get_likelihood(serial_interval, mu, sigma, distn = "normal")
}
\arguments{
\item{serial_interval}{numeric; time between infection or symptom onset of two individuals}

\item{mu}{numeric; mean of serial interval distribution}

\item{sigma}{numeric; standard deviation of serial interval distribution}

\item{distn}{string; distribution to be assumed for serial interval. Defaults to "normal", but also accepts "gamma".}
}
\value{
value of the likelihood of the input serial_interval
}
\description{
Two distributions can be specified for the serial interval distribution: Normal and Gamma. A Normal distribution with mean \mu and standard deviation \sigma is assumed by default. If a Gamma distribution is specified, the mean \mu and standard deviation \sigma are converted to the shape and rate parameters for a Gamma distribution as:
[
\beta = \frac{\mu}{\sigma^2}
\alpha = (\frac{\mu}{\beta})^2
]
}
