% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_likelihood.R
\name{get_likelihood}
\alias{get_likelihood}
\title{Compute the likelihood of a transmission pair given the serial interval distribution}
\usage{
get_likelihood(serial_intervals, mu, sigma, distn = "normal")
}
\arguments{
\item{serial_intervals}{numeric vector of serial intervals}

\item{mu}{numeric; mean of serial interval distribution}

\item{sigma}{numeric; standard deviation of serial interval distribution}

\item{distn}{string; distribution to be assumed for serial interval. Defaults to "normal", but also accepts "gamma".}
}
\value{
Likelihood of the transmission pair.
}
\description{
Two distributions can be specified for the serial interval distribution: Normal and Gamma.
A Normal distribution with mean \eqn{\mu} and standard deviation \eqn{\sigma} is assumed by default.
If a Gamma distribution is specified, the mean \eqn{\mu} and standard deviation \eqn{\sigma}
are converted to the shape and rate parameters for a Gamma distribution as:
\deqn{
\beta = \frac{\mu}{\sigma^2}
\alpha = \left(\frac{\mu}{\beta}\right)^2
}
}
