% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_component.R
\name{integrate_component}
\alias{integrate_component}
\title{Integrate Component Function}
\usage{
integrate_component(
  d,
  mu,
  sigma,
  comp,
  dist = c("normal", "gamma"),
  lower = TRUE
)
}
\arguments{
\item{d}{numeric; the data point.}

\item{mu}{numeric; the mean value.}

\item{sigma}{numeric, the standard deviation.}

\item{comp}{integer; the component number, accepts values 1 to 7.}

\item{dist}{string; assumed distribution of the serial interval; take "normal" or "gamma".}

\item{lower}{logical; indicates whether to integrate the lower part (TRUE) or the upper part (FALSE).}
}
\value{
The integrated value.
}
\description{
This function integrates the component function over a specified interval.
}
