% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_components_wrapper.R
\name{integrate_components_wrapper}
\alias{integrate_components_wrapper}
\title{Wrapper function to for integrate_components()}
\usage{
integrate_components_wrapper(d, mu, sigma, dist = "normal")
}
\arguments{
\item{d}{numeric; the data point.}

\item{mu}{numeric; the mean value.}

\item{sigma}{numeric; the standard deviation.}

\item{dist}{string; assumed distribution of the serial interval; takes "normal" or "gamma"; defaults to "normal"}
}
\value{
A vector containing the integrated values for each component.
}
\description{
This function computes the integral for all components for a given data point.
}
\examples{
# Example 1: Compute integrations for a specific data point
integrate_components_wrapper(1, 10, 2)
}
