% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_components_wrapper.R
\name{integrate_components_wrapper}
\alias{integrate_components_wrapper}
\title{Wrapper function to for integrate_components()}
\usage{
integrate_components_wrapper(d, mu, sigma)
}
\arguments{
\item{d}{numeric; the data point.}

\item{mu}{numeric; the mean value.}

\item{sigma}{numeric; the standard deviation.}
}
\value{
A vector containing the integrated values for each component.
The components are:
\itemize{
\item Component 1: Integration for CP route.
\item Component 2: Integration for PS route (mean = mu, sd = sigma).
\item Component 3: Integration for PS route (mean = -mu, sd = sigma).
\item Component 4: Integration for PT route (mean = 2*mu, sd = sqrt(2)*sigma).
\item Component 5: Integration for PT route (mean = -2*mu, sd = sqrt(2)*sigma).
\item Component 6: Integration for PQ route (mean = 3*mu, sd = sqrt(3)*sigma).
\item Component 7: Integration for PQ route (mean = -3*mu, sd = sqrt(3)*sigma).
}
}
\description{
This function computes the integral for all components for a given data point.
}
\examples{
# Example 1: Compute integrations for a specific data point
integrate_components_wrapper(1, 10, 2)
}
