% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_si_fit.R
\name{plot_si_fit}
\alias{plot_si_fit}
\title{Plot the epidemic curve and fitted serial interval distribution}
\usage{
plot_si_fit(dat, mean, sd, weights, dist = "normal")
}
\arguments{
\item{dat}{data frame containing a variable x with index case to case (ICC) intervals}

\item{mean}{mean of serial interval distribution}

\item{sd}{standard deviation of serial interval distribution}

\item{weights}{numeric vector of weights based on transmission route}

\item{dist}{string; assumed distribution of the serial interval; takes "normal" or "gamma"; defaults to "normal".}
}
\value{
ggplot object
}
\description{
Plot the epidemic curve and fitted serial interval distribution
}
