% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_estim.R
\name{rt_estim}
\alias{rt_estim}
\title{Estimate Time-varying Reproduction Number}
\usage{
rt_estim(
  inc_dat,
  mean_si,
  sd_si,
  dist_si = c("normal", "gamma"),
  cut_tail = NULL,
  pos_only = TRUE
)
}
\arguments{
\item{inc_dat}{data frame; data frame with incidence data. The data frame should have two columns: inc (daily incidence) and date.}

\item{mean_si}{numeric; mean of serial interval distribution}

\item{sd_si}{numeric; standard deviation of serial interval distribution}

\item{dist_si}{string; distribution to be assumed for serial interval. Accepts "normal" or "gamma".}

\item{cut_tail}{numeric; number of days beyond which the likelihood of transmission between two events is 0. Defaults to NULL.}

\item{pos_only}{logical; if TRUE, only positive serial intervals are considered. If the serial interval is negative, the function returns 0.}
}
\value{
A numeric vector with the expected reproduction number for each day.
}
\description{
This function estimates the time-varying reproduction number (R_t) using the method proposed by Wallinga and Teunis (AJE 2004).
The function takes as input a data frame with incidence data and requires the specification of the serial interval distribution.
The user must input the mean and standard deviation of the serial interval distribution and select the functional form: Normal or Gamma.
}
