% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_estim.R
\name{si_estim}
\alias{si_estim}
\title{Estimate serial interval using the method from Vink et al. (2014)}
\usage{
si_estim(dat)
}
\arguments{
\item{dat}{a (numeric) vector of index case to case intervals}
}
\value{
vector with estimates for the mean and standard deviation of the primary-secondary infection component
}
\description{
Estimate serial interval using the method from Vink et al. (2014)
}
\examples{
N <- 1000; hmu<-15; hsigma<-3
CP <- rhalfnorm((0.2*N),theta=sqrt(pi/2)/(sqrt(2)*hsigma))
PS <- rnorm(0.5*N,mean=hmu,sd=hsigma)
PT <- rnorm(0.3*N,mean=2*hmu,sd=sqrt(2)*hsigma)
PQ <- rnorm(0.1*N,mean=3*hmu,sd=sqrt(3)*hsigma)
icc_intervals <- round(c(CP,PS,PT,PQ))

si_estim(icc_intervals)
}
