% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_estim.R
\name{si_estim}
\alias{si_estim}
\title{Estimate serial interval using the method from Vink et al. (2014)}
\usage{
si_estim(dat, n = 50)
}
\arguments{
\item{dat}{a (numeric) vector of index case to case intervals}

\item{n}{integer; number of iterations for EM algorithm; defaults to n = 50}
}
\value{
vector with estimates for the mean and standard deviation of the primary-secondary infection component
}
\description{
Estimate serial interval using the method from Vink et al. (2014)
}
\examples{
my_data<-c(rep(1,38),rep(2,39),rep(3,30),rep(4,17),rep(5,7))

si_estim(my_data)
}
