% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_int.R
\name{vec_int}
\alias{vec_int}
\title{Vectorized integration}
\usage{
vec_int(x, l, u, s, m, rt, q)
}
\arguments{
\item{x}{a (numeric) vector of index case to case intervals}

\item{l}{lower bound for integration}

\item{u}{upper bound for integration}

\item{s}{standard deviation of distribution function to be integrated}

\item{m}{mean of distribution function to be integrated}

\item{rt}{route of transmission; integer from 1 to 7}

\item{q}{character string; quantity to estimate, "zero", "lower", "upper"}
}
\value{
a vector of solutions for the integral for every value of x
}
\description{
Vectorized integration
}
\examples{
df <-c(rep(1,38),rep(2,39),rep(3,30),rep(4,17),rep(5,7))

vec_int(x = df, l = df, u = (df + 1), s = sd(df), m = mean(df), rt = 1, q = "zero")
}
